+++ libs/cardano-ledger-pretty/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog for `cardano-ledger-pretty`
 
-## 1.3.2.1
+## 1.3.3.0
 
-*
+* Add `PrettyA` instance for `NoUpdate`
 
 ## 1.3.2.0